+++ affise_attribution_module_adservice/ios/affise_attribution_module_adservice/Sources/affise_attribution_module_adservice/AffiseAttributionModuleAdservicePlugin.swift
@@ -1,0 +1,6 @@
+import Flutter
+import UIKit
+
+public class AffiseAttributionModuleAdservicePlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {}
+}