+++ affise_attribution_module_advertising/ios/Classes/AffiseAttributionModuleAdvertisingPlugin.swift
@@ -1,5 +1,0 @@
-import Flutter
-
-public class AffiseAttributionModuleAdvertisingPlugin: NSObject, FlutterPlugin {
-  public static func register(with registrar: FlutterPluginRegistrar) {}
-}