+++ affise_attribution_module_androidid/android/src/main/kotlin/com/affise/affise_attribution_module_androidid/AffiseAttributionModuleAndroididPlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_androidid
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleAndroididPlugin */
+class AffiseAttributionModuleAndroididPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}