+++ affise_attribution_module_appsflyer/android/src/main/kotlin/com/affise/affise_attribution_module_appsflyer/AffiseAttributionModuleAppsflyerPlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_appsflyer
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleAppsflyerPlugin */
+class AffiseAttributionModuleAppsflyerPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}