+++ affise_attribution_module_google/ios/affise_attribution_module_google/Sources/affise_attribution_module_google/AffiseAttributionModuleGooglePlugin.swift
@@ -1,0 +1,6 @@
+import Flutter
+import UIKit
+
+public class AffiseAttributionModuleGooglePlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {}
+}