+++ affise_attribution_module_huawei/android/src/main/kotlin/com/affise/affise_attribution_module_huawei/AffiseAttributionModuleHuaweiPlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_huawei
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleHuaweiPlugin */
+class AffiseAttributionModuleHuaweiPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}