+++ affise_attribution_module_link/android/src/main/kotlin/com/affise/affise_attribution_module_link/AffiseAttributionModuleLinkPlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_link
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleLinkPlugin */
+class AffiseAttributionModuleLinkPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}