+++ affise_attribution_module_link/example/android/app/src/main/kotlin/com/affise/affise_attribution_module_link_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.affise.affise_attribution_module_link_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()