+++ affise_attribution_module_meta/android/src/main/kotlin/com/affise/affise_attribution_module_meta/AffiseAttributionModuleMetaPlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_meta
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleMetaPlugin */
+class AffiseAttributionModuleMetaPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}