+++ affise_attribution_module_network/android/src/main/kotlin/com/affise/affise_attribution_module_network/AffiseAttributionModuleNetworkPlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_network
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleNetworkPlugin */
+class AffiseAttributionModuleNetworkPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}