+++ affise_attribution_module_persistent/ios/affise_attribution_module_persistent/Sources/affise_attribution_module_persistent/AffiseAttributionModulePersistentPlugin.swift
@@ -1,0 +1,6 @@
+import Flutter
+import UIKit
+
+public class AffiseAttributionModulePersistentPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {}
+}