+++ affise_attribution_module_phone/android/src/main/kotlin/com/affise/affise_attribution_module_phone/AffiseAttributionModulePhonePlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_phone
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModulePhonePlugin */
+class AffiseAttributionModulePhonePlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}