+++ affise_attribution_module_rustore/android/src/main/kotlin/com/affise/affise_attribution_module_rustore/AffiseAttributionModuleRustorePlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_rustore
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleRustorePlugin */
+class AffiseAttributionModuleRustorePlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}