+++ affise_attribution_module_status/android/src/main/kotlin/com/affise/affise_attribution_module_status/AffiseAttributionModuleStatusPlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_status
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleStatusPlugin */
+class AffiseAttributionModuleStatusPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}