+++ affise_attribution_module_subscription/android/src/main/kotlin/com/affise/affise_attribution_module_subscription/AffiseAttributionModuleSubscriptionPlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_subscription
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleSubscriptionPlugin */
+class AffiseAttributionModuleSubscriptionPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}