+++ affise_attribution_module_subscription/ios/affise_attribution_module_subscription/Sources/affise_attribution_module_subscription/AffiseAttributionModuleSubscriptionPlugin.swift
@@ -1,0 +1,6 @@
+import Flutter
+import UIKit
+
+public class AffiseAttributionModuleSubscriptionPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {}
+}