+++ affise_attribution_module_tiktok/android/src/main/kotlin/com/affise/affise_attribution_module_tiktok/AffiseAttributionModuleTiktokPlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_tiktok
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleTiktokPlugin */
+class AffiseAttributionModuleTiktokPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}