+++ affise_attribution_module_tiktok/ios/affise_attribution_module_tiktok/Sources/affise_attribution_module_tiktok/AffiseAttributionModuleTiktokPlugin.swift
@@ -1,0 +1,6 @@
+import Flutter
+import UIKit
+
+public class AffiseAttributionModuleTiktokPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {}
+}