+++ affise_attribution_module_advertising/CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
+## [1.7.15] - 2026-09-11
+
+### Changed
+
+- Update native iOS to [`1.7.15`](https://github.com/affise/affise-mmp-sdk-ios/blob/1.7.15/CHANGELOG.md)
+- Update native Android to [`1.7.15`](https://github.com/affise/affise-mmp-sdk-android/blob/v1.7.15/CHANGELOG.md)
+
 ## [1.7.14] - 2026-09-01
 
 ### Added
 
-- Flutter package for native Affise module `Advertising` 
+- Flutter package for native Affise module `affise_attribution_module_advertising` 
 - Advertising module wrapper `com.affise:module-advertising` for Android.
 - Advertising module wrapper `AffiseModule/Advertising` for iOS.
 
@@ -13,4 +20,5 @@
 - Update native iOS to [`1.7.14`](https://github.com/affise/affise-mmp-sdk-ios/blob/1.7.14/CHANGELOG.md)
 - Update native Android to [`1.7.14`](https://github.com/affise/affise-mmp-sdk-android/blob/v1.7.14/CHANGELOG.md)
 
+[1.7.15]: https://github.com/affise/affise-mmp-sdk-flutter/compare/1.7.14...1.7.15
 [1.7.14]: https://github.com/affise/affise-mmp-sdk-flutter/compare/1.7.13...1.7.14